% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{addin_build_project}
\alias{addin_build_project}
\title{RStudio addin: build the current Quone project}
\usage{
addin_build_project()
}
\description{
RStudio addin: build the current Quone project
}
\keyword{internal}
