% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{addin_fmt_active}
\alias{addin_fmt_active}
\title{RStudio addin: format the active editor file in place}
\usage{
addin_fmt_active()
}
\description{
RStudio addin: format the active editor file in place
}
\keyword{internal}
