% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{addin_fmt_check_active}
\alias{addin_fmt_check_active}
\title{RStudio addin: run \code{fmt --check} on the active editor file}
\usage{
addin_fmt_check_active()
}
\description{
RStudio addin: run \code{fmt --check} on the active editor file
}
\keyword{internal}
