% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{confirm}
\alias{confirm}
\title{Tidyverse-style yes/no prompt that degrades gracefully when
\code{usethis} is missing or we're in a non-interactive session.}
\usage{
confirm(message, default = TRUE)
}
\description{
Tidyverse-style yes/no prompt that degrades gracefully when
\code{usethis} is missing or we're in a non-interactive session.
}
\keyword{internal}
