% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{detect_arch}
\alias{detect_arch}
\title{Normalised host arch string used in release-asset names}
\usage{
detect_arch()
}
\description{
Normalised host arch string used in release-asset names
}
\keyword{internal}
