% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_project}
\alias{fmt_project}
\title{Recursively format every \code{.Q} file under a project's \verb{src/}}
\usage{
fmt_project(project = ".", check = FALSE)
}
\arguments{
\item{project}{Project root.}

\item{check}{Run in check mode.}
}
\value{
Invisibly, \code{TRUE} if every file was already (or became)
canonical.
}
\description{
Recursively format every \code{.Q} file under a project's \verb{src/}
}
