% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{format_diagnostic}
\alias{format_diagnostic}
\title{Format a single diagnostic as a one-line \code{cli}-styled string}
\usage{
format_diagnostic(d)
}
\arguments{
\item{d}{A single-row diagnostics tibble or a list with the same
fields.}
}
\value{
A character string.
}
\description{
Format a single diagnostic as a one-line \code{cli}-styled string
}
