% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{persist_compiler_path}
\alias{persist_compiler_path}
\title{Append \verb{QUONEC=<path>} to the user-level \code{.Renviron}, replacing any
previous quone-managed entry. Returns \code{TRUE} on success.}
\usage{
persist_compiler_path(dest)
}
\description{
Append \verb{QUONEC=<path>} to the user-level \code{.Renviron}, replacing any
previous quone-managed entry. Returns \code{TRUE} on success.
}
\keyword{internal}
