% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{print_setup_summary}
\alias{print_setup_summary}
\title{Print the closing "what's next?" summary.}
\usage{
print_setup_summary(compiler, lsp)
}
\description{
Print the closing "what's next?" summary.
}
\keyword{internal}
