% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{remove_dependency}
\alias{remove_dependency}
\title{Remove a dependency from \code{quone.toml}}
\usage{
remove_dependency(package, project = ".")
}
\arguments{
\item{package}{R package name to remove.}

\item{project}{Project root.}
}
\value{
Invisibly, the updated \code{quone.toml} lines.
}
\description{
Remove a dependency from \code{quone.toml}
}
