% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{use_fmt_pre_commit}
\alias{use_fmt_pre_commit}
\title{Install a git pre-commit hook that runs \verb{quonec fmt --check}}
\usage{
use_fmt_pre_commit(project = ".")
}
\arguments{
\item{project}{Project root.}
}
\value{
Invisibly, the path to the new hook.
}
\description{
Install a git pre-commit hook that runs \verb{quonec fmt --check}
}
