% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{user_compiler_path}
\alias{user_compiler_path}
\title{Path that \code{install_compiler()} writes into.}
\usage{
user_compiler_path()
}
\description{
Path that \code{install_compiler()} writes into.
}
\keyword{internal}
