% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{warn_if_not_terminal}
\alias{warn_if_not_terminal}
\title{Print a friendly heads-up if \code{quone::repl()} is being launched from
a context where stdin is unlikely to reach the spawned compiler
(most IDE R consoles).}
\usage{
warn_if_not_terminal()
}
\description{
Print a friendly heads-up if \code{quone::repl()} is being launched from
a context where stdin is unlikely to reach the spawned compiler
(most IDE R consoles).
}
\keyword{internal}
