% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{watch_stop}
\alias{watch_stop}
\title{Stop a watcher}
\usage{
watch_stop(handle = NULL)
}
\arguments{
\item{handle}{An optional handle returned from \code{\link[=watch]{watch()}}.}
}
\value{
Invisibly, \code{TRUE} if a watcher was stopped.
}
\description{
Stops the most recently started watcher, or the specific \code{handle}
if one is provided. Subsequent calls are no-ops.
}
